+++ packages/@jsii/go-runtime/.gitignore
@@ -1,9 +1,9 @@
 /jsii-calc/
-/jsii-runtime-go/embedded/resources/
+/jsii-runtime-go/internal/embedded/resources/
 
 *.generated.*
 
 *.js
 *.d.ts
 
-!jsii-runtime-go/kernel/process/jsii-mock-runtime.js
+!jsii-runtime-go/internal/kernel/process/jsii-mock-runtime.js
