+++ packages/@jsii/go-runtime/build-tools/gen.ts
@@ -14,7 +14,12 @@ const EMBEDDED_RUNTIME_SOURCE_ROOT = resolve(
 
 const RUNTIME_ROOT = resolve(__dirname, '..', 'jsii-runtime-go');
 
-const EMBEDDED_RESOURCE_DIR = join(RUNTIME_ROOT, 'embedded', 'resources');
+const EMBEDDED_RESOURCE_DIR = join(
+  RUNTIME_ROOT,
+  'internal',
+  'embedded',
+  'resources',
+);
 
 mkdirpSync(EMBEDDED_RESOURCE_DIR);
 copySync(EMBEDDED_RUNTIME_SOURCE_ROOT, EMBEDDED_RESOURCE_DIR, {
@@ -25,7 +30,7 @@ copySync(EMBEDDED_RUNTIME_SOURCE_ROOT, EMBEDDED_RESOURCE_DIR, {
   recursive: true,
 });
 
-const KERNEL_LIB_DIR = resolve(RUNTIME_ROOT, 'kernel');
+const KERNEL_LIB_DIR = resolve(RUNTIME_ROOT, 'internal', 'kernel');
 const code = new CodeMaker({ indentationLevel: 1, indentCharacter: '\t' });
 
 const VERSION_FILE = 'version.generated.go';
