+++ packages/@jsii/go-runtime/jsii-runtime-go/internal/embedded/doc.go
