+++ packages/@jsii/go-runtime/jsii-runtime-go/internal/embedded/embedded.go
