+++ packages/@jsii/go-runtime/jsii-runtime-go/internal/kernel/begin.go
@@ -1,7 +1,7 @@
 package kernel
 
 import (
-	"github.com/aws/jsii-runtime-go/api"
+	"github.com/aws/jsii-runtime-go/internal/api"
 )
 
 type BeginProps struct {
