+++ packages/@jsii/go-runtime/jsii-runtime-go/internal/kernel/callbacks.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"reflect"
 
-	"github.com/aws/jsii-runtime-go/api"
+	"github.com/aws/jsii-runtime-go/internal/api"
 )
 
 type callback struct {
