+++ packages/@jsii/go-runtime/jsii-runtime-go/internal/kernel/conversions.go
@@ -3,7 +3,7 @@ package kernel
 import (
 	"reflect"
 
-	"github.com/aws/jsii-runtime-go/api"
+	"github.com/aws/jsii-runtime-go/internal/api"
 )
 
 // CastAndSetToPtr accepts a pointer to any type and attempts to cast the value
