+++ packages/@jsii/go-runtime/jsii-runtime-go/internal/kernel/create.go
@@ -1,6 +1,6 @@
 package kernel
 
-import "github.com/aws/jsii-runtime-go/api"
+import "github.com/aws/jsii-runtime-go/internal/api"
 
 type CreateProps struct {
 	FQN        api.FQN        `json:"fqn"`
