+++ packages/@jsii/go-runtime/jsii-runtime-go/internal/kernel/del.go
@@ -1,6 +1,6 @@
 package kernel
 
-import "github.com/aws/jsii-runtime-go/api"
+import "github.com/aws/jsii-runtime-go/internal/api"
 
 type DelProps struct {
 	ObjRef api.ObjectRef `json:"objref"`
