+++ packages/@jsii/go-runtime/jsii-runtime-go/internal/kernel/doc.go
