+++ packages/@jsii/go-runtime/jsii-runtime-go/internal/kernel/invoke.go
@@ -1,6 +1,6 @@
 package kernel
 
-import "github.com/aws/jsii-runtime-go/api"
+import "github.com/aws/jsii-runtime-go/internal/api"
 
 type InvokeProps struct {
 	Method    string        `json:"method"`
