+++ packages/@jsii/go-runtime/jsii-runtime-go/internal/kernel/process/consume-stderr.go
