+++ packages/@jsii/go-runtime/jsii-runtime-go/internal/kernel/process/jsii-mock-runtime.js
