+++ packages/@jsii/go-runtime/jsii-runtime-go/internal/objectstore/doc.go
