+++ packages/@jsii/go-runtime/jsii-runtime-go/internal/typeregistry/discover-implemented.go
@@ -3,7 +3,7 @@ package typeregistry
 import (
 	"reflect"
 
-	"github.com/aws/jsii-runtime-go/api"
+	"github.com/aws/jsii-runtime-go/internal/api"
 )
 
 // DiscoverImplementation determines the list of registered interfaces that are
