+++ packages/@jsii/go-runtime/jsii-runtime-go/internal/typeregistry/doc.go
