+++ packages/@jsii/go-runtime/jsii-runtime-go/internal/typeregistry/registration.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"reflect"
 
-	"github.com/aws/jsii-runtime-go/api"
+	"github.com/aws/jsii-runtime-go/internal/api"
 )
 
 type typeKind uint8
