+++ packages/@jsii/go-runtime/jsii-runtime-go/internal/typeregistry/type-registry.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"reflect"
 
-	"github.com/aws/jsii-runtime-go/api"
+	"github.com/aws/jsii-runtime-go/internal/api"
 )
 
 // TypeRegistry is used to record runtime type information about the loaded
